Car updated to measure floating point current itself. Fixed a bug where current value was always being set to zero.

+++ FRDM-TFC/Project_Headers/sensors/current.h
@@ -14,7 +14,7 @@
 typedef struct {
 	uint16_t data[CURRENT_FILTER_BUFFER_SIZE];
 	rbuf_uint16_s buffer;
-	int16_t value;
+	float value;
 } MotorCurrent_s;
 extern MotorCurrent_s MotorCurrent[NUM_MOTORS];
